libpam: fix possible heap overflow in _pam_strdup
It is possible to trigger an integer overflow in _pam_strdup
if the passed string is longer than INT_MAX, which could lead
to a smaller memory allocation than needed for the strcpy call.

This in turn could lead to a heap overflow.

+++ libpam/pam_misc.c
@@ -123,7 +123,7 @@ char *_pam_strdup(const char *x)
      register char *new=NULL;
 
      if (x != NULL) {
-	  register int len;
+	  register size_t len;
 
 	  len = strlen (x) + 1;  /* length of string including NUL */
 	  if ((new = malloc(len)) == NULL) {
